+++ src/Models/index.ts
@@ -1,4 +1,5 @@
 export { GetClientListResponse } from './getClientListResponse';
 export { ClientList } from './clientList';
+export { DeviceAction } from './deviceAction';
 export { SingleClientData } from './singleClientData';
 export { Client } from './client';