+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   Placeholder for the next version (at the beginning of the line):
   ## **WORK IN PROGRESS**
 -->
+## **WORK IN PROGRESS**
+* (theimo1221) Implement Device Actions reboot and reconnect
+
 ## 0.0.2-alpha.1 (2022-07-10)
 * (theimo1221) Refactoring
 * (theomo1221) Implement Basic Client Models